GridNearLockRequestV2 example implementation.
