+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/distributed/near/GridNearLockRequestV2.java
@@ -1,0 +1,583 @@
+package org.apache.ignite.internal.processors.cache.distributed.near;
+
+import org.apache.ignite.IgniteCheckedException;
+import org.apache.ignite.internal.processors.affinity.AffinityTopologyVersion;
+import org.apache.ignite.internal.processors.cache.CacheEntryPredicate;
+import org.apache.ignite.internal.processors.cache.GridCacheContext;
+import org.apache.ignite.internal.processors.cache.GridCacheSharedContext;
+import org.apache.ignite.internal.processors.cache.KeyCacheObject;
+import org.apache.ignite.internal.processors.cache.distributed.GridDistributedLockRequest;
+import org.apache.ignite.internal.processors.cache.version.GridCacheVersion;
+import org.apache.ignite.internal.util.tostring.GridToStringInclude;
+import org.apache.ignite.internal.util.typedef.internal.S;
+import org.apache.ignite.lang.IgniteUuid;
+import org.apache.ignite.plugin.extensions.communication.MessageCollectionItemType;
+import org.apache.ignite.plugin.extensions.communication.MessageReader;
+import org.apache.ignite.plugin.extensions.communication.MessageWriter;
+import org.apache.ignite.transactions.TransactionIsolation;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+import java.io.Externalizable;
+import java.nio.ByteBuffer;
+import java.util.Arrays;
+import java.util.UUID;
+
+public class GridNearLockRequestV2 extends GridDistributedLockRequest implements GridNearLockRequest {
+    /** */
+    private static final long serialVersionUID = 0L;
+
+    public static final UUID DUMMY_UUID = new UUID(0, 0);
+
+    /** Topology version. */
+    private AffinityTopologyVersion topVer;
+
+    /** Mini future ID. */
+    private int miniId;
+
+    /** Filter. */
+    private CacheEntryPredicate[] filter;
+
+    /** Implicit flag. */
+    private boolean implicitTx;
+
+    /** Implicit transaction with one key flag. */
+    private boolean implicitSingleTx;
+
+    /** Flag is kept for backward compatibility. */
+    private boolean onePhaseCommit;
+
+    /** Array of mapped DHT versions for this entry. */
+    @GridToStringInclude
+    private GridCacheVersion[] dhtVers;
+
+    /** Subject ID. */
+    private UUID subjId;
+
+    /** Task name hash. */
+    private int taskNameHash;
+
+    /** Has transforms flag. */
+    private boolean hasTransforms;
+
+    /** Sync commit flag. */
+    private boolean syncCommit;
+
+    /** TTL for read operation. */
+    private long accessTtl;
+
+    /** Flag indicating whether cache operation requires a previous value. */
+    private boolean retVal;
+
+    /** {@code True} if first lock request for lock operation sent from client node. */
+    private boolean firstClientReq;
+
+    /**
+     * Empty constructor required for {@link Externalizable}.
+     */
+    public GridNearLockRequestV2() {
+        // No-op.
+    }
+
+    /**
+     * @param cacheId Cache ID.
+     * @param topVer Topology version.
+     * @param nodeId Node ID.
+     * @param threadId Thread ID.
+     * @param futId Future ID.
+     * @param lockVer Cache version.
+     * @param isInTx {@code True} if implicit transaction lock.
+     * @param implicitTx Flag to indicate that transaction is implicit.
+     * @param implicitSingleTx Implicit-transaction-with-one-key flag.
+     * @param isRead Indicates whether implicit lock is for read or write operation.
+     * @param retVal Return value flag.
+     * @param isolation Transaction isolation.
+     * @param isInvalidate Invalidation flag.
+     * @param timeout Lock timeout.
+     * @param keyCnt Number of keys.
+     * @param txSize Expected transaction size.
+     * @param syncCommit Synchronous commit flag.
+     * @param subjId Subject ID.
+     * @param taskNameHash Task name hash code.
+     * @param accessTtl TTL for read operation.
+     * @param skipStore Skip store flag.
+     * @param firstClientReq {@code True} if first lock request for lock operation sent from client node.
+     * @param addDepInfo Deployment info flag.
+     */
+    public GridNearLockRequestV2(
+        int cacheId,
+        @NotNull AffinityTopologyVersion topVer,
+        UUID nodeId,
+        long threadId,
+        IgniteUuid futId,
+        GridCacheVersion lockVer,
+        boolean isInTx,
+        boolean implicitTx,
+        boolean implicitSingleTx,
+        boolean isRead,
+        boolean retVal,
+        TransactionIsolation isolation,
+        boolean isInvalidate,
+        long timeout,
+        int keyCnt,
+        int txSize,
+        boolean syncCommit,
+        @Nullable UUID subjId,
+        int taskNameHash,
+        long accessTtl,
+        boolean skipStore,
+        boolean keepBinary,
+        boolean firstClientReq,
+        boolean addDepInfo
+
+    ) {
+        super(
+            cacheId,
+            nodeId,
+            lockVer,
+            threadId,
+            futId,
+            lockVer,
+            isInTx,
+            isRead,
+            isolation,
+            isInvalidate,
+            timeout,
+            keyCnt,
+            txSize,
+            skipStore,
+            keepBinary,
+            addDepInfo);
+
+        assert topVer.compareTo(AffinityTopologyVersion.ZERO) > 0;
+
+        this.topVer = topVer;
+        this.implicitTx = implicitTx;
+        this.implicitSingleTx = implicitSingleTx;
+        this.syncCommit = syncCommit;
+        this.subjId = subjId;
+        this.taskNameHash = taskNameHash;
+        this.accessTtl = accessTtl;
+        this.retVal = retVal;
+        this.firstClientReq = firstClientReq;
+
+        dhtVers = new GridCacheVersion[keyCnt];
+    }
+
+    /**
+     * @return {@code True} if first lock request for lock operation sent from client node.
+     */
+    public boolean firstClientRequest() {
+        return firstClientReq;
+    }
+
+    /**
+     * @return Topology version.
+     */
+    @Override public AffinityTopologyVersion topologyVersion() {
+        return topVer;
+    }
+
+    /**
+     * @return Subject ID.
+     */
+    public UUID subjectId() {
+        return subjId;
+    }
+
+    /**
+     * @return Task name hash.q
+     */
+    public int taskNameHash() {
+        return taskNameHash;
+    }
+
+    /**
+     * @return Implicit transaction flag.
+     */
+    public boolean implicitTx() {
+        return implicitTx;
+    }
+
+    /**
+     * @return Implicit-transaction-with-one-key flag.
+     */
+    public boolean implicitSingleTx() {
+        return implicitSingleTx;
+    }
+
+    /**
+     * @return Sync commit flag.
+     */
+    public boolean syncCommit() {
+        return syncCommit;
+    }
+
+    /**
+     * @return Filter.
+     */
+    public CacheEntryPredicate[] filter() {
+        return filter;
+    }
+
+    /**
+     * @param filter Filter.
+     * @param ctx Context.
+     * @throws IgniteCheckedException If failed.
+     */
+    public void filter(CacheEntryPredicate[] filter, GridCacheContext ctx)
+        throws IgniteCheckedException {
+        this.filter = filter;
+    }
+
+    /**
+     * @return Mini future ID.
+     */
+    public int miniId() {
+        return miniId;
+    }
+
+    /**
+     * @param miniId Mini future Id.
+     */
+    public void miniId(int miniId) {
+        this.miniId = miniId;
+    }
+
+    public IgniteUuid oldVersionMiniId() {
+        return new IgniteUuid(DUMMY_UUID, miniId);
+    }
+
+    /**
+     * @param hasTransforms {@code True} if originating transaction has transform entries.
+     */
+    public void hasTransforms(boolean hasTransforms) {
+        this.hasTransforms = hasTransforms;
+    }
+
+    /**
+     * @return {@code True} if originating transaction has transform entries.
+     */
+    public boolean hasTransforms() {
+        return hasTransforms;
+    }
+
+    /**
+     * @return Need return value flag.
+     */
+    public boolean needReturnValue() {
+        return retVal;
+    }
+
+    /**
+     * Adds a key.
+     *
+     * @param key Key.
+     * @param retVal Flag indicating whether value should be returned.
+     * @param dhtVer DHT version.
+     * @param ctx Context.
+     * @throws IgniteCheckedException If failed.
+     */
+    public void addKeyBytes(
+        KeyCacheObject key,
+        boolean retVal,
+        @Nullable GridCacheVersion dhtVer,
+        GridCacheContext ctx
+    ) throws IgniteCheckedException {
+        dhtVers[idx] = dhtVer;
+
+        // Delegate to super.
+        addKeyBytes(key, retVal, ctx);
+    }
+
+    /**
+     * @param idx Index of the key.
+     * @return DHT version for key at given index.
+     */
+    public GridCacheVersion dhtVersion(int idx) {
+        return dhtVers[idx];
+    }
+
+    /**
+     * @return TTL for read operation.
+     */
+    public long accessTtl() {
+        return accessTtl;
+    }
+
+    /** {@inheritDoc} */
+    @Override public void prepareMarshal(GridCacheSharedContext ctx) throws IgniteCheckedException {
+        super.prepareMarshal(ctx);
+
+        if (filter != null) {
+            GridCacheContext cctx = ctx.cacheContext(cacheId);
+
+            for (CacheEntryPredicate p : filter) {
+                if (p != null)
+                    p.prepareMarshal(cctx);
+            }
+        }
+    }
+
+    /** {@inheritDoc} */
+    @Override public void finishUnmarshal(GridCacheSharedContext ctx, ClassLoader ldr) throws IgniteCheckedException {
+        super.finishUnmarshal(ctx, ldr);
+
+        if (filter != null) {
+            GridCacheContext cctx = ctx.cacheContext(cacheId);
+
+            for (CacheEntryPredicate p : filter) {
+                if (p != null)
+                    p.finishUnmarshal(cctx, ldr);
+            }
+        }
+    }
+
+    /** {@inheritDoc} */
+    @Override public boolean writeTo(ByteBuffer buf, MessageWriter writer) {
+        writer.setBuffer(buf);
+
+        if (!super.writeTo(buf, writer))
+            return false;
+
+        if (!writer.isHeaderWritten()) {
+            if (!writer.writeHeader(directType(), fieldsCount()))
+                return false;
+
+            writer.onHeaderWritten();
+        }
+
+        switch (writer.state()) {
+            case 20:
+                if (!writer.writeLong("accessTtl", accessTtl))
+                    return false;
+
+                writer.incrementState();
+
+            case 21:
+                if (!writer.writeObjectArray("dhtVers", dhtVers, MessageCollectionItemType.MSG))
+                    return false;
+
+                writer.incrementState();
+
+            case 22:
+                if (!writer.writeObjectArray("filter", filter, MessageCollectionItemType.MSG))
+                    return false;
+
+                writer.incrementState();
+
+            case 23:
+                if (!writer.writeBoolean("firstClientReq", firstClientReq))
+                    return false;
+
+                writer.incrementState();
+
+            case 24:
+                if (!writer.writeBoolean("hasTransforms", hasTransforms))
+                    return false;
+
+                writer.incrementState();
+
+            case 25:
+                if (!writer.writeBoolean("implicitSingleTx", implicitSingleTx))
+                    return false;
+
+                writer.incrementState();
+
+            case 26:
+                if (!writer.writeBoolean("implicitTx", implicitTx))
+                    return false;
+
+                writer.incrementState();
+
+            case 27:
+                if (!writer.writeInt("miniId", miniId))
+                    return false;
+
+                writer.incrementState();
+
+            case 28:
+                if (!writer.writeBoolean("onePhaseCommit", onePhaseCommit))
+                    return false;
+
+                writer.incrementState();
+
+            case 29:
+                if (!writer.writeBoolean("retVal", retVal))
+                    return false;
+
+                writer.incrementState();
+
+            case 30:
+                if (!writer.writeUuid("subjId", subjId))
+                    return false;
+
+                writer.incrementState();
+
+            case 31:
+                if (!writer.writeBoolean("syncCommit", syncCommit))
+                    return false;
+
+                writer.incrementState();
+
+            case 32:
+                if (!writer.writeInt("taskNameHash", taskNameHash))
+                    return false;
+
+                writer.incrementState();
+
+            case 33:
+                if (!writer.writeMessage("topVer", topVer))
+                    return false;
+
+                writer.incrementState();
+
+        }
+
+        return true;
+    }
+
+    /** {@inheritDoc} */
+    @Override public boolean readFrom(ByteBuffer buf, MessageReader reader) {
+        reader.setBuffer(buf);
+
+        if (!reader.beforeMessageRead())
+            return false;
+
+        if (!super.readFrom(buf, reader))
+            return false;
+
+        switch (reader.state()) {
+            case 20:
+                accessTtl = reader.readLong("accessTtl");
+
+                if (!reader.isLastRead())
+                    return false;
+
+                reader.incrementState();
+
+            case 21:
+                dhtVers = reader.readObjectArray("dhtVers", MessageCollectionItemType.MSG, GridCacheVersion.class);
+
+                if (!reader.isLastRead())
+                    return false;
+
+                reader.incrementState();
+
+            case 22:
+                filter = reader.readObjectArray("filter", MessageCollectionItemType.MSG, CacheEntryPredicate.class);
+
+                if (!reader.isLastRead())
+                    return false;
+
+                reader.incrementState();
+
+            case 23:
+                firstClientReq = reader.readBoolean("firstClientReq");
+
+                if (!reader.isLastRead())
+                    return false;
+
+                reader.incrementState();
+
+            case 24:
+                hasTransforms = reader.readBoolean("hasTransforms");
+
+                if (!reader.isLastRead())
+                    return false;
+
+                reader.incrementState();
+
+            case 25:
+                implicitSingleTx = reader.readBoolean("implicitSingleTx");
+
+                if (!reader.isLastRead())
+                    return false;
+
+                reader.incrementState();
+
+            case 26:
+                implicitTx = reader.readBoolean("implicitTx");
+
+                if (!reader.isLastRead())
+                    return false;
+
+                reader.incrementState();
+
+            case 27:
+                miniId = reader.readInt("miniId");
+
+                if (!reader.isLastRead())
+                    return false;
+
+                reader.incrementState();
+
+            case 28:
+                onePhaseCommit = reader.readBoolean("onePhaseCommit");
+
+                if (!reader.isLastRead())
+                    return false;
+
+                reader.incrementState();
+
+            case 29:
+                retVal = reader.readBoolean("retVal");
+
+                if (!reader.isLastRead())
+                    return false;
+
+                reader.incrementState();
+
+            case 30:
+                subjId = reader.readUuid("subjId");
+
+                if (!reader.isLastRead())
+                    return false;
+
+                reader.incrementState();
+
+            case 31:
+                syncCommit = reader.readBoolean("syncCommit");
+
+                if (!reader.isLastRead())
+                    return false;
+
+                reader.incrementState();
+
+            case 32:
+                taskNameHash = reader.readInt("taskNameHash");
+
+                if (!reader.isLastRead())
+                    return false;
+
+                reader.incrementState();
+
+            case 33:
+                topVer = reader.readMessage("topVer");
+
+                if (!reader.isLastRead())
+                    return false;
+
+                reader.incrementState();
+
+        }
+
+        return reader.afterMessageRead(GridNearLockRequestV1.class);
+    }
+
+    /** {@inheritDoc} */
+    @Override public byte directType() {
+        return 125;
+    }
+
+    /** {@inheritDoc} */
+    @Override public byte fieldsCount() {
+        return 34;
+    }
+
+    /** {@inheritDoc} */
+    @Override public String toString() {
+        return S.toString(GridNearLockRequestV2.class, this, "filter", Arrays.toString(filter),
+            "super", super.toString());
+    }
+}
