+++ modules/core/src/test/java/org/apache/ignite/internal/processors/cache/distributed/IgniteCacheClientNodeChangingTopologyTest.java
@@ -64,7 +64,7 @@ import org.apache.ignite.internal.processors.cache.GridCacheEntryEx;
 import org.apache.ignite.internal.processors.cache.distributed.dht.atomic.GridNearAtomicUpdateRequest;
 import org.apache.ignite.internal.processors.cache.distributed.near.GridNearCacheAdapter;
 import org.apache.ignite.internal.processors.cache.distributed.near.GridNearCacheEntry;
-import org.apache.ignite.internal.processors.cache.distributed.near.GridNearLockRequest;
+import org.apache.ignite.internal.processors.cache.distributed.near.GridNearLockRequestV1;
 import org.apache.ignite.internal.processors.cache.distributed.near.GridNearTxPrepareRequest;
 import org.apache.ignite.internal.processors.cache.version.GridCacheVersion;
 import org.apache.ignite.internal.util.lang.GridAbsPredicate;
@@ -599,10 +599,10 @@ public class IgniteCacheClientNodeChangingTopologyTest extends GridCommonAbstrac
 
         TestCommunicationSpi spi = (TestCommunicationSpi)ignite2.configuration().getCommunicationSpi();
 
-        spi.blockMessages(GridNearLockRequest.class, ignite0.localNode().id());
-        spi.blockMessages(GridNearLockRequest.class, ignite1.localNode().id());
+        spi.blockMessages(GridNearLockRequestV1.class, ignite0.localNode().id());
+        spi.blockMessages(GridNearLockRequestV1.class, ignite1.localNode().id());
 
-        spi.record(GridNearLockRequest.class);
+        spi.record(GridNearLockRequestV1.class);
 
         final IgniteCache<Integer, Integer> cache = ignite2.cache(null);
 
@@ -638,19 +638,19 @@ public class IgniteCacheClientNodeChangingTopologyTest extends GridCommonAbstrac
 
         List<Object> msgs = spi.recordedMessages();
 
-        assertTrue(((GridNearLockRequest)msgs.get(0)).firstClientRequest());
-        assertTrue(((GridNearLockRequest)msgs.get(1)).firstClientRequest());
+        assertTrue(((GridNearLockRequestV1)msgs.get(0)).firstClientRequest());
+        assertTrue(((GridNearLockRequestV1)msgs.get(1)).firstClientRequest());
 
         for (int i = 2; i < msgs.size(); i++)
-            assertFalse(((GridNearLockRequest)msgs.get(i)).firstClientRequest());
+            assertFalse(((GridNearLockRequestV1)msgs.get(i)).firstClientRequest());
 
         ignite3.close();
 
         for (int i = 0; i < 100; i++)
             map.put(i, i + 1);
 
-        spi.blockMessages(GridNearLockRequest.class, ignite0.localNode().id());
-        spi.blockMessages(GridNearLockRequest.class, ignite1.localNode().id());
+        spi.blockMessages(GridNearLockRequestV1.class, ignite0.localNode().id());
+        spi.blockMessages(GridNearLockRequestV1.class, ignite1.localNode().id());
 
         putFut = GridTestUtils.runAsync(new Callable<Object>() {
             @Override public Object call() throws Exception {
@@ -812,9 +812,9 @@ public class IgniteCacheClientNodeChangingTopologyTest extends GridCommonAbstrac
         final Integer key1 = keys.get1();
         final Integer key2 = keys.get2();
 
-        spi.blockMessages(GridNearLockRequest.class, ignite0.localNode().id());
-        spi.blockMessages(GridNearLockRequest.class, ignite1.localNode().id());
-        spi.blockMessages(GridNearLockRequest.class, ignite2.localNode().id());
+        spi.blockMessages(GridNearLockRequestV1.class, ignite0.localNode().id());
+        spi.blockMessages(GridNearLockRequestV1.class, ignite1.localNode().id());
+        spi.blockMessages(GridNearLockRequestV1.class, ignite2.localNode().id());
 
         final IgniteCache<Integer, Integer> cache = ignite3.cache(null);
 
@@ -912,11 +912,11 @@ public class IgniteCacheClientNodeChangingTopologyTest extends GridCommonAbstrac
         for (int i = 0; i < 100; i++)
             map.put(i, i);
 
-        spi.blockMessages(GridNearLockRequest.class, ignite0.localNode().id());
-        spi.blockMessages(GridNearLockRequest.class, ignite1.localNode().id());
-        spi.blockMessages(GridNearLockRequest.class, ignite2.localNode().id());
+        spi.blockMessages(GridNearLockRequestV1.class, ignite0.localNode().id());
+        spi.blockMessages(GridNearLockRequestV1.class, ignite1.localNode().id());
+        spi.blockMessages(GridNearLockRequestV1.class, ignite2.localNode().id());
 
-        spi.record(GridNearLockRequest.class);
+        spi.record(GridNearLockRequestV1.class);
 
         final IgniteCache<Integer, Integer> cache = ignite3.cache(null);
 
@@ -1155,8 +1155,8 @@ public class IgniteCacheClientNodeChangingTopologyTest extends GridCommonAbstrac
 
         TestCommunicationSpi spi = (TestCommunicationSpi)ignite2.configuration().getCommunicationSpi();
 
-        spi.blockMessages(GridNearLockRequest.class, ignite0.localNode().id());
-        spi.blockMessages(GridNearLockRequest.class, ignite1.localNode().id());
+        spi.blockMessages(GridNearLockRequestV1.class, ignite0.localNode().id());
+        spi.blockMessages(GridNearLockRequestV1.class, ignite1.localNode().id());
 
         final IgniteCache<Integer, Integer> cache = ignite2.cache(null);
 
@@ -1263,7 +1263,7 @@ public class IgniteCacheClientNodeChangingTopologyTest extends GridCommonAbstrac
 
         TestCommunicationSpi spi = (TestCommunicationSpi)ignite3.configuration().getCommunicationSpi();
 
-        spi.record(GridNearLockRequest.class);
+        spi.record(GridNearLockRequestV1.class);
 
         IgniteCache<Integer, Integer> cache = ignite3.cache(null);
 
@@ -1302,7 +1302,7 @@ public class IgniteCacheClientNodeChangingTopologyTest extends GridCommonAbstrac
 
         TestCommunicationSpi spi0 = (TestCommunicationSpi)ignite0.configuration().getCommunicationSpi();
 
-        spi0.record(GridNearLockRequest.class);
+        spi0.record(GridNearLockRequestV1.class);
 
         List<Integer> keys = primaryKeys(ignite1.cache(null), 3, 0);
 
@@ -1321,7 +1321,7 @@ public class IgniteCacheClientNodeChangingTopologyTest extends GridCommonAbstrac
         assertEquals(3, msgs.size());
 
         for (Object msg : msgs)
-            assertFalse(((GridNearLockRequest)msg).firstClientRequest());
+            assertFalse(((GridNearLockRequestV1)msg).firstClientRequest());
     }
 
     /**
@@ -1331,10 +1331,10 @@ public class IgniteCacheClientNodeChangingTopologyTest extends GridCommonAbstrac
     private void checkClientLockMessages(List<Object> msgs, int expCnt) {
         assertEquals(expCnt, msgs.size());
 
-        assertTrue(((GridNearLockRequest)msgs.get(0)).firstClientRequest());
+        assertTrue(((GridNearLockRequestV1)msgs.get(0)).firstClientRequest());
 
         for (int i = 1; i < msgs.size(); i++)
-            assertFalse(((GridNearLockRequest)msgs.get(i)).firstClientRequest());
+            assertFalse(((GridNearLockRequestV1)msgs.get(i)).firstClientRequest());
     }
 
     /**
